refactor : Move Parser and Divider Class
- to utility package
